correcao de arquitetura e criacao de metodos proprios de transferencia em classe pai Conta e classes filhas

+++ src/main/java/com/cesurg/projetopagamento/core/domain/usecase/ContaUseCaseImpl.java
@@ -19,7 +19,7 @@ public class ContaUseCaseImpl implements ContaUseCase {
     }
 
     @Override
-    public void atualizarSaldo(String id, Integer saldo) {
+    public void atualizarSaldo(String id, double saldo) {
         contaRepository.atualizarSaldo(id, saldo);
     }
