+++ src/main/java/com/cesurg/projetopagamento/core/interfaces/ContaUseCase.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface ContaUseCase {
     void criarConta (Conta conta);
-    void atualizarSaldo (String id, Integer saldo);
+    void atualizarSaldo (String id, double saldo);
     void deletarConta (String id);
     List<Conta> listarConta ();
 }
